requirements.txt
streamlit>=1.24.0
openai>=1.0.0
langdetect>=1.0.9
pillow>=9.0.0
